use float

+++ SolaxHub/Knx/KnxSolaxProcessor.cs
@@ -74,9 +74,9 @@ namespace SolaxHub.Knx
                 // InverterStatus - 6.020 status with mode
                 yield return UpdateValue(nameof(SolaxData.InverterStatus), new[] { (byte)solaxData.InverterStatus });
                 // YieldToday - 14
-                yield return UpdateValue(nameof(SolaxData.YieldToday), BitConverter.GetBytes((ushort)solaxData.YieldToday));
+                yield return UpdateValue(nameof(SolaxData.YieldToday), BitConverter.GetBytes((float)solaxData.YieldToday));
                 // YieldTotal - 14
-                yield return UpdateValue(nameof(SolaxData.YieldTotal), BitConverter.GetBytes((ushort)solaxData.YieldTotal));
+                yield return UpdateValue(nameof(SolaxData.YieldTotal), BitConverter.GetBytes((float)solaxData.YieldTotal));
             }
         }
